fix: supported tools for claude 2.1
